fix: rename palettes to themes in all the source code
