+++ lib/models/config_model.dart
@@ -54,8 +54,8 @@ class ConfigModel {
   /// Preferred display mode for the system list (e.g., 'grid', 'list').
   final String systemViewMode;
 
-  /// Identifier of the currently active UI palette.
-  final String paletteName;
+  /// Identifier of the currently active UI theme.
+  final String themeName;
 
   /// Whether to display detailed game metadata by default.
   final bool showGameInfo;
@@ -148,7 +148,7 @@ class ConfigModel {
     this.emulators = const {},
     this.gameViewMode = 'list',
     this.systemViewMode = 'grid',
-    this.paletteName = 'system',
+    this.themeName = 'system',
     this.showGameInfo = false,
     this.isFullscreen = true,
     this.bartopExitPoweroff = false,
@@ -215,7 +215,7 @@ class ConfigModel {
       emulators: emulators,
       gameViewMode: (json['gameViewMode'] ?? 'list').toString(),
       systemViewMode: (json['systemViewMode'] ?? 'grid').toString(),
-      paletteName: (json['paletteName'] ?? 'system').toString(),
+      themeName: (json['themeName'] ?? 'system').toString(),
       showGameInfo:
           (json['showGameInfo'] ?? false).toString().toLowerCase() == 'true',
       isFullscreen:
@@ -335,7 +335,7 @@ class ConfigModel {
       'emulators': emulatorsJson,
       'gameViewMode': gameViewMode,
       'systemViewMode': systemViewMode,
-      'paletteName': paletteName,
+      'themeName': themeName,
       'showGameInfo': showGameInfo,
       'isFullscreen': isFullscreen,
       'bartopExitPoweroff': bartopExitPoweroff,
@@ -373,7 +373,7 @@ class ConfigModel {
     Map<String, EmulatorModel>? emulators,
     String? gameViewMode,
     String? systemViewMode,
-    String? paletteName,
+    String? themeName,
     bool? showGameInfo,
     bool? isFullscreen,
     bool? bartopExitPoweroff,
@@ -408,7 +408,7 @@ class ConfigModel {
       emulators: emulators ?? this.emulators,
       gameViewMode: gameViewMode ?? this.gameViewMode,
       systemViewMode: systemViewMode ?? this.systemViewMode,
-      paletteName: paletteName ?? this.paletteName,
+      themeName: themeName ?? this.themeName,
       showGameInfo: showGameInfo ?? this.showGameInfo,
       isFullscreen: isFullscreen ?? this.isFullscreen,
       bartopExitPoweroff: bartopExitPoweroff ?? this.bartopExitPoweroff,
