+++ lib/themes/app_themes.dart
@@ -1,0 +1,193 @@
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+import 'package:neostation/providers/theme_provider.dart';
+
+// Import all individual themes
+import 'dark_theme.dart' as dark;
+import 'light_theme.dart' as light;
+import 'oled_theme.dart' as oled;
+import 'valentine_theme.dart' as valentine;
+import 'dracula_theme.dart' as dracula;
+import 'nord_theme.dart' as nord;
+import 'coffee_theme.dart' as coffee;
+import 'tokyo_night_theme.dart' as tokyo_night;
+import 'solarized_light_theme.dart' as solarized_light;
+import 'abyss_theme.dart' as abyss;
+import 'cyberpunk_theme.dart' as cyberpunk;
+import 'solarized_dark_theme.dart' as solarized_dark;
+import 'palenight_theme.dart' as palenight;
+import 'horizon_theme.dart' as horizon;
+
+class AppThemes {
+  static String getLogoPath() {
+    return 'assets/images/logo_transparent.png';
+  }
+
+  // References to individual themes
+  static ThemeData get darkTheme => dark.darkTheme;
+  static ThemeData get lightTheme => light.lightTheme;
+  static ThemeData get oledTheme => oled.oledTheme;
+  static ThemeData get valentineTheme => valentine.valentineTheme;
+  static ThemeData get draculaTheme => dracula.draculaTheme;
+  static ThemeData get nordTheme => nord.nordTheme;
+  static ThemeData get coffeeTheme => coffee.coffeeTheme;
+  static ThemeData get tokyoNightTheme => tokyo_night.tokyoNightTheme;
+  static ThemeData get solarizedLightTheme =>
+      solarized_light.solarizedLightTheme;
+  static ThemeData get abyssTheme => abyss.abyssTheme;
+  static ThemeData get cyberpunkTheme => cyberpunk.cyberpunkTheme;
+  static ThemeData get solarizedDarkTheme => solarized_dark.solarizedDarkTheme;
+  static ThemeData get palenightTheme => palenight.palenightTheme;
+  static ThemeData get horizonTheme => horizon.horizonTheme;
+
+  // References to custom colors for each theme
+  static dynamic get darkCustomColors => dark.DarkCustomColors();
+  static dynamic get lightCustomColors => light.LightCustomColors();
+  static dynamic get oledCustomColors => oled.OledCustomColors();
+  static dynamic get valentineCustomColors => valentine.ValentineCustomColors();
+  static dynamic get draculaCustomColors => dracula.DraculaCustomColors();
+  static dynamic get nordCustomColors => nord.NordCustomColors();
+  static dynamic get coffeeCustomColors => coffee.CoffeeCustomColors();
+  static dynamic get tokyoNightCustomColors =>
+      tokyo_night.TokyoNightCustomColors();
+  static dynamic get solarizedLightCustomColors =>
+      solarized_light.SolarizedLightCustomColors();
+  static dynamic get abyssCustomColors => abyss.AbyssCustomColors();
+  static dynamic get cyberpunkCustomColors => cyberpunk.CyberpunkCustomColors();
+  static dynamic get solarizedDarkCustomColors =>
+      solarized_dark.SolarizedDarkCustomColors();
+  static dynamic get palenightCustomColors => palenight.PalenightCustomColors();
+  static dynamic get horizonCustomColors => horizon.HorizonCustomColors();
+
+  /// Retrieves header colors based on the current context's theme.
+  static dynamic getCustomColors(BuildContext context) {
+    // Prefer detection by theme name if a ThemeProvider is available (more reliable).
+    try {
+      final themeProvider = Provider.of<ThemeProvider>(context, listen: false);
+      final themeName = themeProvider.currentThemeName;
+      String resolvedThemeName = themeName;
+
+      if (themeName == 'system') {
+        final brightness =
+            WidgetsBinding.instance.platformDispatcher.platformBrightness;
+        resolvedThemeName = brightness == Brightness.dark ? 'dark' : 'light';
+      }
+
+      switch (resolvedThemeName) {
+        case 'light':
+          return light.LightCustomColors();
+        case 'oled':
+          return oled.OledCustomColors();
+        case 'valentine':
+          return valentine.ValentineCustomColors();
+        case 'dracula':
+          return dracula.DraculaCustomColors();
+        case 'nord':
+          return nord.NordCustomColors();
+        case 'coffee':
+          return coffee.CoffeeCustomColors();
+        case 'tokyo_night':
+          return tokyo_night.TokyoNightCustomColors();
+        case 'solarized_light':
+          return solarized_light.SolarizedLightCustomColors();
+        case 'abyss':
+          return abyss.AbyssCustomColors();
+        case 'cyberpunk':
+          return cyberpunk.CyberpunkCustomColors();
+        case 'solarized_dark':
+          return solarized_dark.SolarizedDarkCustomColors();
+        case 'palenight':
+          return palenight.PalenightCustomColors();
+        case 'horizon':
+          return horizon.HorizonCustomColors();
+        default:
+          return dark.DarkCustomColors();
+      }
+    } catch (_) {
+      // Fallback to color comparison if provider is not available.
+    }
+
+    final scheme = Theme.of(context).colorScheme;
+    final surface = scheme.surface;
+    final secondary = scheme.secondary;
+
+    // Compare with surface or secondary colors of each theme.
+    if (surface == lightTheme.colorScheme.surface ||
+        secondary == lightTheme.colorScheme.secondary) {
+      return light.LightCustomColors();
+    } else if (surface == oledTheme.colorScheme.surface ||
+        secondary == oledTheme.colorScheme.secondary) {
+      return oled.OledCustomColors();
+    } else if (surface == valentineTheme.colorScheme.surface ||
+        secondary == valentineTheme.colorScheme.secondary) {
+      return valentine.ValentineCustomColors();
+    } else if (surface == draculaTheme.colorScheme.surface ||
+        secondary == draculaTheme.colorScheme.secondary) {
+      return dracula.DraculaCustomColors();
+    } else if (surface == nordTheme.colorScheme.surface ||
+        secondary == nordTheme.colorScheme.secondary) {
+      return nord.NordCustomColors();
+    } else if (surface == coffeeTheme.colorScheme.surface ||
+        secondary == coffeeTheme.colorScheme.secondary) {
+      return coffee.CoffeeCustomColors();
+    } else if (surface == tokyoNightTheme.colorScheme.surface ||
+        secondary == tokyoNightTheme.colorScheme.secondary) {
+      return tokyo_night.TokyoNightCustomColors();
+    } else if (surface == solarizedLightTheme.colorScheme.surface ||
+        secondary == solarizedLightTheme.colorScheme.secondary) {
+      return solarized_light.SolarizedLightCustomColors();
+    } else if (surface == abyssTheme.colorScheme.surface ||
+        secondary == abyssTheme.colorScheme.secondary) {
+      return abyss.AbyssCustomColors();
+    } else if (surface == cyberpunkTheme.colorScheme.surface ||
+        secondary == cyberpunkTheme.colorScheme.secondary) {
+      return cyberpunk.CyberpunkCustomColors();
+    } else if (surface == solarizedDarkTheme.colorScheme.surface ||
+        secondary == solarizedDarkTheme.colorScheme.secondary) {
+      return solarized_dark.SolarizedDarkCustomColors();
+    } else if (surface == palenightTheme.colorScheme.surface ||
+        secondary == palenightTheme.colorScheme.secondary) {
+      return palenight.PalenightCustomColors();
+    } else if (surface == horizonTheme.colorScheme.surface ||
+        secondary == horizonTheme.colorScheme.secondary) {
+      return horizon.HorizonCustomColors();
+    } else {
+      return dark.DarkCustomColors();
+    }
+  }
+
+  static ThemeData getThemeDataByName(String themeName) {
+    switch (themeName) {
+      case 'dark':
+        return darkTheme;
+      case 'light':
+        return lightTheme;
+      case 'oled':
+        return oledTheme;
+      case 'valentine':
+        return valentineTheme;
+      case 'dracula':
+        return draculaTheme;
+      case 'nord':
+        return nordTheme;
+      case 'coffee':
+        return coffeeTheme;
+      case 'tokyo_night':
+        return tokyoNightTheme;
+      case 'solarized_light':
+        return solarizedLightTheme;
+      case 'abyss':
+        return abyssTheme;
+      case 'cyberpunk':
+        return cyberpunkTheme;
+      case 'solarized_dark':
+        return solarizedDarkTheme;
+      case 'palenight':
+        return palenightTheme;
+      case 'horizon':
+        return horizonTheme;
+      default:
+        return darkTheme;
+    }
+  }
+}
