+++ lib/themes/dark_theme.dart
@@ -1,44 +1,49 @@
 import 'package:flutter/material.dart';
 import 'package:neostation/themes/corner_radii.dart';
 
-const Color _primaryColor = Color(0xFF4078F2);
-const Color _onPrimaryColor = Color(0xFFFFFFFF);
-const Color _secondaryColor = Color(0xFF50A14F);
-const Color _onSecondaryColor = Color(0xFFFFFFFF);
-const Color _tertiaryColor = Color(0xFFC18401);
-const Color _onTertiaryColor = Color(0xFFFFFFFF);
-const Color _surfaceColor = Color(0xFFE5E5E6);
-const Color _onSurfaceColor = Color(0xFF383A42);
-const Color _errorColor = Color(0xFFE45649);
-const Color _onErrorColor = Color(0xFFFFFFFF);
-const Color _outlineColor = Color(0xFFA0A1A7);
-const Color _shadowColor = Color(0x1A000000);
+const Color _primaryColor = Color(0xFF605dff);
+const Color _onPrimaryColor = Color(0xFFedf1fe);
+const Color _secondaryColor = Color(0xFFf43098);
+const Color _onSecondaryColor = Color(0xFFf9e4f0);
+const Color _tertiaryColor = Color(0xFF00d3bb);
+const Color _onTertiaryColor = Color(0xFF084d49);
+const Color _tertiaryFixedColor = Color(0xFF09090b);
+const Color _onTertiaryFixedColor = Color(0xFFe4e4e7);
+const Color _surfaceColor = Color(0xFF1d232a);
+const Color _onSurfaceColor = Color(0xFFecf9ff);
 
-const Color _backgroundColor = Color(0xFFFAFAFA);
+const Color _outlineColor = Color(0xFF323841);
+const Color _shadowColor = Color(0xFF09090b);
 
-const Color _batteryFull = Color(0xFF50A14F);
-const Color _batteryMedium = Color(0xFFC18401);
-const Color _batteryLow = Color(0xFFE45649);
-const Color _batteryPower = Color(0xFF4078F2);
+const Color _backgroundColor = Color(0xFF15191e);
 
-const Color _warningColor = Color(0xFFC18401);
-const Color _onWarningColor = Color(0xFFFFFFFF);
-const Color _successColor = Color(0xFF50A14F);
-const Color _onSuccessColor = Color(0xFFFFFFFF);
-const Color _infoColor = Color(0xFF4078F2);
-const Color _onInfoColor = Color(0xFFFFFFFF);
+const Color _batteryFull = Color(0xFF00d390);
+const Color _batteryMedium = Color(0xFFfcb700);
+const Color _batteryLow = Color(0xFFff627d);
+const Color _batteryPower = Color(0xFF00bafe);
 
-final ThemeData oneLightPalette = ThemeData(
+const Color _errorColor = Color(0xFFff627d);
+const Color _onErrorColor = Color(0xFF4d0218);
+const Color _warningColor = Color(0xFFfcb700);
+const Color _onWarningColor = Color(0xFF793205);
+const Color _successColor = Color(0xFF00d390);
+const Color _onSuccessColor = Color(0xFF004c39);
+const Color _infoColor = Color(0xFF00bafe);
+const Color _onInfoColor = Color(0xFF042e49);
+
+final ThemeData darkTheme = ThemeData(
   useMaterial3: true,
-  colorScheme: ColorScheme.light(
+  colorScheme: ColorScheme.dark(
     primary: _primaryColor,
     secondary: _secondaryColor,
     tertiary: _tertiaryColor,
+    tertiaryFixed: _tertiaryFixedColor,
     surface: _surfaceColor,
 
     onPrimary: _onPrimaryColor,
     onSecondary: _onSecondaryColor,
     onTertiary: _onTertiaryColor,
+    onTertiaryFixed: _onTertiaryFixedColor,
     onSurface: _onSurfaceColor,
 
     error: _errorColor,
@@ -49,7 +54,7 @@ final ThemeData oneLightPalette = ThemeData(
 
   cardColor: _backgroundColor,
   scaffoldBackgroundColor: _backgroundColor,
-  extensions: [CornerRadii.m()],
+  extensions: [CornerRadii.sm()],
 
   textTheme: TextTheme(
     displayLarge: TextStyle(
@@ -80,7 +85,7 @@ final ThemeData oneLightPalette = ThemeData(
   ),
 );
 
-class OneLightCustomColors {
+class DarkCustomColors {
   Color get batteryFull => _batteryFull;
   Color get batteryMedium => _batteryMedium;
   Color get batteryLow => _batteryLow;
