+++ lib/themes/nord_theme.dart
@@ -1,44 +1,49 @@
 import 'package:flutter/material.dart';
 import 'package:neostation/themes/corner_radii.dart';
 
-const Color _primaryColor = Color(0xFF268BD2);
-const Color _onPrimaryColor = Color(0xFFFDF6E3);
-const Color _secondaryColor = Color(0xFF2AA198);
-const Color _onSecondaryColor = Color(0xFFFDF6E3);
-const Color _tertiaryColor = Color(0xFF859900);
-const Color _onTertiaryColor = Color(0xFFFDF6E3);
-const Color _surfaceColor = Color(0xFFEEE8D5);
-const Color _onSurfaceColor = Color(0xFF657B83);
-const Color _errorColor = Color(0xFFDC322F);
-const Color _onErrorColor = Color(0xFFFDF6E3);
-const Color _outlineColor = Color(0xFF93A1A1);
-const Color _shadowColor = Color(0x1A000000);
+const Color _primaryColor = Color(0xFF5e81ac);
+const Color _onPrimaryColor = Color(0xFF03060b);
+const Color _secondaryColor = Color(0xFF81a1c1);
+const Color _onSecondaryColor = Color(0xFF06090d);
+const Color _tertiaryColor = Color(0xFF88c0d0);
+const Color _onTertiaryColor = Color(0xFF070d10);
+const Color _tertiaryFixedColor = Color(0xFF4c566a);
+const Color _onTertiaryFixedColor = Color(0xFFd8dee9);
+const Color _surfaceColor = Color(0xFFeceff4);
+const Color _onSurfaceColor = Color(0xFF2e3440);
 
-const Color _backgroundColor = Color(0xFFFDF6E3);
+const Color _outlineColor = Color(0xFFd8dee9);
+const Color _shadowColor = Color(0xFF2e3440);
 
-const Color _batteryFull = Color(0xFF859900);
-const Color _batteryMedium = Color(0xFFB58900);
-const Color _batteryLow = Color(0xFFDC322F);
-const Color _batteryPower = Color(0xFF268BD2);
+const Color _backgroundColor = Color(0xFFd8dee9);
 
-const Color _warningColor = Color(0xFFB58900);
-const Color _onWarningColor = Color(0xFFFDF6E3);
-const Color _successColor = Color(0xFF859900);
-const Color _onSuccessColor = Color(0xFFFDF6E3);
-const Color _infoColor = Color(0xFF268BD2);
-const Color _onInfoColor = Color(0xFFFDF6E3);
+const Color _batteryFull = Color(0xFFa3be8d);
+const Color _batteryMedium = Color(0xFFebcb8b);
+const Color _batteryLow = Color(0xFFbf616a);
+const Color _batteryPower = Color(0xFFb48ead);
 
-final ThemeData solarizedLightPalette = ThemeData(
+const Color _errorColor = Color(0xFFbf616a);
+const Color _onErrorColor = Color(0xFF0d0304);
+const Color _warningColor = Color(0xFFebcb8b);
+const Color _onWarningColor = Color(0xFF130f07);
+const Color _successColor = Color(0xFFa3be8d);
+const Color _onSuccessColor = Color(0xFF0a0d07);
+const Color _infoColor = Color(0xFFb48ead);
+const Color _onInfoColor = Color(0xFF0c070b);
+
+final ThemeData nordTheme = ThemeData(
   useMaterial3: true,
-  colorScheme: ColorScheme.light(
+  colorScheme: ColorScheme.dark(
     primary: _primaryColor,
     secondary: _secondaryColor,
     tertiary: _tertiaryColor,
+    tertiaryFixed: _tertiaryFixedColor,
     surface: _surfaceColor,
 
     onPrimary: _onPrimaryColor,
     onSecondary: _onSecondaryColor,
     onTertiary: _onTertiaryColor,
+    onTertiaryFixed: _onTertiaryFixedColor,
     onSurface: _onSurfaceColor,
 
     error: _errorColor,
@@ -80,7 +85,7 @@ final ThemeData solarizedLightPalette = ThemeData(
   ),
 );
 
-class SolarizedLightCustomColors {
+class NordCustomColors {
   Color get batteryFull => _batteryFull;
   Color get batteryMedium => _batteryMedium;
   Color get batteryLow => _batteryLow;
@@ -88,13 +93,10 @@ class SolarizedLightCustomColors {
 
   Color get errorColor => _errorColor;
   Color get onErrorColor => _onErrorColor;
-
   Color get successColor => _successColor;
   Color get onSuccessColor => _onSuccessColor;
-
   Color get infoColor => _infoColor;
   Color get onInfoColor => _onInfoColor;
-
   Color get warningColor => _warningColor;
   Color get onWarningColor => _onWarningColor;
 }
