+++ lib/themes/solarized_dark_theme.dart
@@ -1,49 +1,46 @@
 import 'package:flutter/material.dart';
 import 'package:neostation/themes/corner_radii.dart';
 
-const Color _primaryColor = Color(0xFFf43098);
-const Color _onPrimaryColor = Color(0xFFffffff);
-const Color _secondaryColor = Color(0xFFab44ff);
-const Color _onSecondaryColor = Color(0xFFf8f3fd);
-const Color _tertiaryColor = Color(0xFF71d1fe);
-const Color _onTertiaryColor = Color(0xFF014a70);
-const Color _tertiaryFixedColor = Color(0xFF830c41);
-const Color _onTertiaryFixedColor = Color(0xFFf9cbe5);
-const Color _surfaceColor = Color(0xFFfcf2f7);
-const Color _onSurfaceColor = Color(0xFFc5005a);
+const Color _primaryColor = Color(0xFF268BD2);
+const Color _onPrimaryColor = Color(0xFF002B36);
+const Color _secondaryColor = Color(0xFF2AA198);
+const Color _onSecondaryColor = Color(0xFF002B36);
+const Color _tertiaryColor = Color(0xFF859900);
+const Color _onTertiaryColor = Color(0xFF002B36);
+final Color _surfaceColor = HSLColor.fromColor(
+  const Color(0xFF002B36),
+).withLightness(0.25).toColor();
+const Color _onSurfaceColor = Color(0xFF839496);
+const Color _errorColor = Color(0xFFDC322F);
+const Color _onErrorColor = Color(0xFF002B36);
+const Color _outlineColor = Color(0xFF586E75);
+const Color _shadowColor = Color(0xFF000000);
 
-const Color _outlineColor = Color(0xFFf9cbe5);
-const Color _shadowColor = Color(0xFF830c41);
+const Color _backgroundColor = Color(0xFF002B36);
 
-const Color _backgroundColor = Color(0xFFf9cbe5);
+const Color _batteryFull = Color(0xFF859900);
+const Color _batteryMedium = Color(0xFFB58900);
+const Color _batteryLow = Color(0xFFDC322F);
+const Color _batteryPower = Color(0xFF268BD2);
 
-const Color _batteryFull = Color(0xFF5ce8b3);
-const Color _batteryMedium = Color(0xFFff8904);
-const Color _batteryLow = Color(0xFFf82834);
-const Color _batteryPower = Color(0xFF51e8fb);
+const Color _warningColor = Color(0xFFB58900);
+const Color _onWarningColor = Color(0xFF002B36);
+const Color _successColor = Color(0xFF859900);
+const Color _onSuccessColor = Color(0xFF002B36);
+const Color _infoColor = Color(0xFF268BD2);
+const Color _onInfoColor = Color(0xFF002B36);
 
-const Color _errorColor = Color(0xFFf82834);
-const Color _onErrorColor = Color(0xFFfef2f2);
-const Color _warningColor = Color(0xFFff8904);
-const Color _onWarningColor = Color(0xFF421104);
-const Color _successColor = Color(0xFF5ce8b3);
-const Color _onSuccessColor = Color(0xFF006044);
-const Color _infoColor = Color(0xFF51e8fb);
-const Color _onInfoColor = Color(0xFF005889);
-
-final ThemeData valentinePalette = ThemeData(
+final ThemeData solarizedDarkTheme = ThemeData(
   useMaterial3: true,
-  colorScheme: ColorScheme.light(
+  colorScheme: ColorScheme.dark(
     primary: _primaryColor,
     secondary: _secondaryColor,
     tertiary: _tertiaryColor,
-    tertiaryFixed: _tertiaryFixedColor,
     surface: _surfaceColor,
 
     onPrimary: _onPrimaryColor,
     onSecondary: _onSecondaryColor,
     onTertiary: _onTertiaryColor,
-    onTertiaryFixed: _onTertiaryFixedColor,
     onSurface: _onSurfaceColor,
 
     error: _errorColor,
@@ -85,7 +82,7 @@ final ThemeData valentinePalette = ThemeData(
   ),
 );
 
-class ValentineCustomColors {
+class SolarizedDarkCustomColors {
   Color get batteryFull => _batteryFull;
   Color get batteryMedium => _batteryMedium;
   Color get batteryLow => _batteryLow;
@@ -93,10 +90,13 @@ class ValentineCustomColors {
 
   Color get errorColor => _errorColor;
   Color get onErrorColor => _onErrorColor;
+
   Color get successColor => _successColor;
   Color get onSuccessColor => _onSuccessColor;
+
   Color get infoColor => _infoColor;
   Color get onInfoColor => _onInfoColor;
+
   Color get warningColor => _warningColor;
   Color get onWarningColor => _onWarningColor;
 }
