+++ lib/themes/tokyo_night_theme.dart
@@ -1,49 +1,44 @@
 import 'package:flutter/material.dart';
 import 'package:neostation/themes/corner_radii.dart';
 
-const Color _primaryColor = Color(0xFFff79c6);
-const Color _onPrimaryColor = Color(0xFF16050e);
-const Color _secondaryColor = Color(0xFFbd93f9);
-const Color _onSecondaryColor = Color(0xFF0d0815);
-const Color _tertiaryColor = Color(0xFFffb86c);
-const Color _onTertiaryColor = Color(0xFF160d04);
-const Color _tertiaryFixedColor = Color(0xFF414558);
-const Color _onTertiaryFixedColor = Color(0xFFd6d7db);
-const Color _surfaceColor = Color(0xFF282a36);
-const Color _onSurfaceColor = Color(0xFFf8f8f3);
+const Color _primaryColor = Color(0xFF7AA2F7);
+const Color _onPrimaryColor = Color(0xFF1A1B26);
+const Color _secondaryColor = Color(0xFF9ECE6A);
+const Color _onSecondaryColor = Color(0xFF1A1B26);
+const Color _tertiaryColor = Color(0xFF7DCFFF);
+const Color _onTertiaryColor = Color(0xFF1A1B26);
+const Color _surfaceColor = Color(0xFF24283B);
+const Color _onSurfaceColor = Color(0xFFC0CAF5);
+const Color _errorColor = Color(0xFFF7768E);
+const Color _onErrorColor = Color(0xFF1A1B26);
+const Color _outlineColor = Color(0xFF414868);
+const Color _shadowColor = Color(0xFF000000);
 
-const Color _outlineColor = Color(0xFF32343f);
-const Color _shadowColor = Color(0xFF1f202a);
+const Color _backgroundColor = Color(0xFF1A1B26);
 
-const Color _backgroundColor = Color(0xFF1f202a);
+const Color _batteryFull = Color(0xFF9ECE6A);
+const Color _batteryMedium = Color(0xFFE0AF68);
+const Color _batteryLow = Color(0xFFF7768E);
+const Color _batteryPower = Color(0xFF7DCFFF);
 
-const Color _batteryFull = Color(0xFF51fa7b);
-const Color _batteryMedium = Color(0xFFf1fa8c);
-const Color _batteryLow = Color(0xFFff5555);
-const Color _batteryPower = Color(0xFF8be9fd);
+const Color _warningColor = Color(0xFFE0AF68);
+const Color _onWarningColor = Color(0xFF1A1B26);
+const Color _successColor = Color(0xFF9ECE6A);
+const Color _onSuccessColor = Color(0xFF1A1B26);
+const Color _infoColor = Color(0xFF7DCFFF);
+const Color _onInfoColor = Color(0xFF1A1B26);
 
-const Color _errorColor = Color(0xFFff5555);
-const Color _onErrorColor = Color(0xFF160202);
-const Color _warningColor = Color(0xFFf1fa8c);
-const Color _onWarningColor = Color(0xFF141507);
-const Color _successColor = Color(0xFF51fa7b);
-const Color _onSuccessColor = Color(0xFF021505);
-const Color _infoColor = Color(0xFF8be9fd);
-const Color _onInfoColor = Color(0xFF071316);
-
-final ThemeData draculaPalette = ThemeData(
+final ThemeData tokyoNightTheme = ThemeData(
   useMaterial3: true,
   colorScheme: ColorScheme.dark(
     primary: _primaryColor,
     secondary: _secondaryColor,
     tertiary: _tertiaryColor,
-    tertiaryFixed: _tertiaryFixedColor,
     surface: _surfaceColor,
 
     onPrimary: _onPrimaryColor,
     onSecondary: _onSecondaryColor,
     onTertiary: _onTertiaryColor,
-    onTertiaryFixed: _onTertiaryFixedColor,
     onSurface: _onSurfaceColor,
 
     error: _errorColor,
@@ -54,7 +49,7 @@ final ThemeData draculaPalette = ThemeData(
 
   cardColor: _backgroundColor,
   scaffoldBackgroundColor: _backgroundColor,
-  extensions: [CornerRadii.xl()],
+  extensions: [CornerRadii.m()],
 
   textTheme: TextTheme(
     displayLarge: TextStyle(
@@ -85,7 +80,7 @@ final ThemeData draculaPalette = ThemeData(
   ),
 );
 
-class DraculaCustomColors {
+class TokyoNightCustomColors {
   Color get batteryFull => _batteryFull;
   Color get batteryMedium => _batteryMedium;
   Color get batteryLow => _batteryLow;
@@ -93,10 +88,13 @@ class DraculaCustomColors {
 
   Color get errorColor => _errorColor;
   Color get onErrorColor => _onErrorColor;
+
   Color get successColor => _successColor;
   Color get onSuccessColor => _onSuccessColor;
+
   Color get infoColor => _infoColor;
   Color get onInfoColor => _onInfoColor;
+
   Color get warningColor => _warningColor;
   Color get onWarningColor => _onWarningColor;
 }
